refactor: utilModule 분리
- #189
