+++ src/data-calculator/data-calculator.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { DateCalculator } from './date-calculator.component';
+
+@Module({
+  providers: [DateCalculator],
+  exports: [DateCalculator],
+})
+export class DateCalculatorModule {}
