+++ src/ext/ext.module.ts
@@ -1,10 +1,8 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { AuthModule } from 'src/auth/auth.module';
 import { DeviceInfo } from 'src/entities/device-info.entity';
 import { TagLog } from 'src/entities/tag-log.entity';
 import { TagLogRepository } from 'src/tag-log-v1/repository/mysql/tag-log.repository';
-import { TagLogModule } from 'src/tag-log-v1/tag-log.module';
 import { UserModule } from 'src/user/user.module';
 import { ExtService } from './ext.service';
 import { DeviceInfoRepository } from './repository/mysql/device-info.repository';
@@ -24,12 +22,7 @@ const deviceInfoRepo = {
  * 24Hane 기능 외 다른 서비스에서 사용하는 외부 API 모듈
  */
 @Module({
-  imports: [
-    AuthModule,
-    UserModule,
-    TagLogModule,
-    TypeOrmModule.forFeature([DeviceInfo, TagLog]),
-  ],
+  imports: [UserModule, TypeOrmModule.forFeature([DeviceInfo, TagLog])],
   controllers: [Where42Controller],
   providers: [tagLogRepo, deviceInfoRepo, ExtService],
 })
