+++ src/google-api/google-api.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { GoogleApi } from './google-api.component';
+
+@Module({
+  providers: [GoogleApi],
+  exports: [GoogleApi],
+})
+export class GoogleModule {}
