+++ src/tag-log-v1/tag-log.module.ts
@@ -1,10 +1,9 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { AuthModule } from 'src/auth/auth.module';
+import { DateCalculatorModule } from 'src/data-calculator/data-calculator.module';
 import { PairInfo } from 'src/entities/pair-info.entity';
 import { TagLog } from 'src/entities/tag-log.entity';
 import { UserModule } from 'src/user/user.module';
-import { UtilsModule } from 'src/utils/utils.module';
 import { PairInfoRepository } from './repository/mysql/pair-info.repository';
 import { TagLogRepository } from './repository/mysql/tag-log.repository';
 import { TagLogAdminController } from './tag-log-admin.controller';
@@ -24,9 +23,8 @@ const pairInfoRepo = {
 
 @Module({
   imports: [
-    AuthModule,
     TypeOrmModule.forFeature([TagLog, PairInfo]),
-    UtilsModule,
+    DateCalculatorModule,
     UserModule,
   ],
   exports: [TypeOrmModule],
