+++ src/tag-log-v2/tag-log-v2.module.ts
@@ -1,19 +1,17 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { AuthModule } from 'src/auth/auth.module';
+import { DateCalculatorModule } from 'src/data-calculator/data-calculator.module';
 import { DeviceInfo } from 'src/entities/device-info.entity';
 import { PairInfo } from 'src/entities/pair-info.entity';
 import { TagLog } from 'src/entities/tag-log.entity';
 import { StatisticsModule } from 'src/statistics/statictics.module';
 import { UserModule } from 'src/user/user.module';
-import { UtilsModule } from 'src/utils/utils.module';
 import { Cabi42Controller } from '../ext/cabi42.controller';
 import { DeviceInfoRepository } from './repository/mysql/device-info.repository';
 import { PairInfoRepository } from './repository/mysql/pair-info.repository';
 import { TagLogRepository } from './repository/mysql/tag-log.repository';
 import { TagLogAdminController } from './tag-log-v2-admin.controller';
 import { TagLogAdminService } from './tag-log-v2-admin.service';
-
 import { TagLogController } from './tag-log-v2.controller';
 import { TagLogService } from './tag-log-v2.service';
 
@@ -34,9 +32,8 @@ const deviceInfoRepo = {
 
 @Module({
   imports: [
-    AuthModule,
     TypeOrmModule.forFeature([TagLog, PairInfo, DeviceInfo]),
-    UtilsModule,
+    DateCalculatorModule,
     UserModule,
     StatisticsModule,
   ],
